+++ packages/hardhat-polkadot/sample-projects/typescript/hardhat.config.ts
@@ -6,7 +6,9 @@ const config: HardhatUserConfig = {
     solidity: "0.8.28",
     networks: {
         hardhat: {
-            polkadot: true,
+            polkadot: {
+                target: "evm",
+            },
             nodeConfig: {
                 nodeBinaryPath: "./bin/dev-node",
                 rpcPort: 8000,